Move the whole unit tests into separate file "tests.rs"
